requirements/base.txt
beautifulsoup4==4.3.2
Django==1.11.9
Pillow>=3.1.1
django-filebrowser>=3.7.2
django-grappelli>=2.9.1
django-adminactions
django-tagging==0.4.5
django-material==1.1.1
html2text==3.200.3
python-dateutil==2.4.2
six==1.10
vobject==0.6.6
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.5
ipython
django-extensions
easy-thumbnails==2.4.1
factory_boy
coverage
